feat(sendy): delete api + fixes
